feat: add IdentityToFreezeDoesNotExistError for consensus validation
