+++ packages/rs-dpp/src/errors/consensus/state/state_error.rs
@@ -37,6 +37,7 @@ use crate::consensus::state::document::document_not_for_sale_error::DocumentNotF
 use crate::consensus::state::group::{GroupActionAlreadyCompletedError, GroupActionAlreadySignedByIdentityError, GroupActionDoesNotExistError, IdentityMemberOfGroupNotFoundError, IdentityNotMemberOfGroupError, ModificationOfGroupActionMainParametersNotPermittedError};
 use crate::consensus::state::identity::identity_for_token_configuration_not_found_error::IdentityInTokenConfigurationNotFoundError;
 use crate::consensus::state::identity::identity_public_key_already_exists_for_unique_contract_bounds_error::IdentityPublicKeyAlreadyExistsForUniqueContractBoundsError;
+use crate::consensus::state::identity::identity_to_freeze_does_not_exist_error::IdentityToFreezeDoesNotExistError;
 use crate::consensus::state::identity::invalid_identity_contract_nonce_error::InvalidIdentityNonceError;
 use crate::consensus::state::identity::missing_transfer_key_error::MissingTransferKeyError;
 use crate::consensus::state::identity::no_transfer_key_for_core_withdrawal_available_error::NoTransferKeyForCoreWithdrawalAvailableError;
@@ -310,6 +311,9 @@ pub enum StateError {
     ModificationOfGroupActionMainParametersNotPermittedError(
         ModificationOfGroupActionMainParametersNotPermittedError,
     ),
+
+    #[error(transparent)]
+    IdentityToFreezeDoesNotExistError(IdentityToFreezeDoesNotExistError),
 }
 
 impl From<StateError> for ConsensusError {
