+++ packages/rs-drive-abci/src/execution/validation/state_transition/state_transitions/batch/action_validation/token/token_freeze_transition_action/state_v0/mod.rs
@@ -18,11 +18,13 @@ use crate::execution::types::execution_operation::ValidationOperation;
 use crate::execution::types::state_transition_execution_context::StateTransitionExecutionContextMethodsV0;
 use dpp::consensus::ConsensusError;
 use dpp::consensus::state::group::ModificationOfGroupActionMainParametersNotPermittedError;
+use dpp::consensus::state::identity::identity_to_freeze_does_not_exist_error::IdentityToFreezeDoesNotExistError;
 use dpp::group::action_event::GroupActionEvent;
 use dpp::group::group_action::GroupActionAccessors;
 use dpp::tokens::info::v0::IdentityTokenInfoV0Accessors;
 use dpp::tokens::token_event::TokenEvent;
 use drive::state_transition_action::batch::batched_transition::token_transition::token_base_transition_action::TokenBaseTransitionActionAccessorsV0;
+use crate::execution::validation::state_transition::common::validate_identity_exists::validate_identity_exists;
 
 pub(in crate::execution::validation::state_transition::state_transitions::batch::action_validation) trait TokenFreezeTransitionActionStateValidationV0 {
     fn validate_state_v0(
@@ -136,7 +138,24 @@ impl TokenFreezeTransitionActionStateValidationV0 for TokenFreezeTransitionActio
                     )),
                 ));
             }
-        };
+        } else {
+            //make sure the identity to freeze exists if we didn't find info
+            let recipient_exists = validate_identity_exists(
+                platform.drive,
+                &self.identity_to_freeze_id(),
+                execution_context,
+                transaction,
+                platform_version,
+            )?;
+
+            if !recipient_exists {
+                return Ok(SimpleConsensusValidationResult::new_with_error(
+                    ConsensusError::StateError(StateError::IdentityToFreezeDoesNotExistError(
+                        IdentityToFreezeDoesNotExistError::new(self.identity_to_freeze_id()),
+                    )),
+                ));
+            }
+        }
 
         Ok(SimpleConsensusValidationResult::new())
     }
