+++ packages/rs-platform-value/src/types/identifier.rs
@@ -1,15 +1,16 @@
 use bincode::enc::Encoder;
 use bincode::error::EncodeError;
 use bincode::{Decode, Encode};
+use rand::distributions::Standard;
+use rand::prelude::Distribution;
 use rand::rngs::StdRng;
 use rand::Rng;
-use std::convert::{TryFrom, TryInto};
-use std::fmt;
-
 use serde::de::Visitor;
 use serde::{Deserialize, Serialize};
 #[cfg(feature = "json")]
 use serde_json::Value as JsonValue;
+use std::convert::{TryFrom, TryInto};
+use std::fmt;
 
 use crate::string_encoding::{Encoding, ALL_ENCODINGS};
 use crate::types::encoding_string_to_encoding;
@@ -36,6 +37,13 @@ pub struct IdentifierBytes32(pub [u8; 32]);
     Decode,
 )]
 pub struct Identifier(pub IdentifierBytes32);
+
+impl Distribution<Identifier> for Standard {
+    fn sample<R: Rng + ?Sized>(&self, rng: &mut R) -> Identifier {
+        let bytes: [u8; 32] = rng.gen();
+        Identifier::new(bytes)
+    }
+}
 
 impl platform_serialization::PlatformVersionEncode for Identifier {
     fn platform_encode<E: Encoder>(
